+++ test/function/relational/smaller.test.js
@@ -126,18 +126,19 @@ describe('smaller', function() {
     assert.throws(function () {smaller(unit('100cm'), bignumber(22));});
   });
 
-  it('should perform lexical comparison on two strings', function() {
+  it('should compare two strings by their numerical value', function() {
     assert.equal(smaller('0', 0), false);
-    assert.equal(smaller('abd', 'abc'), false);
-    assert.equal(smaller('abc', 'abc'), false);
-    assert.equal(smaller('abc', 'abd'), true);
+    assert.equal(smaller('10', '2'), false);
+    assert.equal(smaller('1e3', '1000'), false);
+
+    assert.throws(function () {smaller('A', 'B')}, /Cannot convert "A" to a number/);
   });
 
   describe('Array', function () {
 
     it('should compare array - scalar', function () {
-      assert.deepEqual(smaller('B', ['A', 'B', 'C']), [false, false, true]);
-      assert.deepEqual(smaller(['A', 'B', 'C'], 'B'), [true, false, false]);
+      assert.deepEqual(smaller(2, [1, 2, 3]), [false, false, true]);
+      assert.deepEqual(smaller([1, 2, 3], 2), [true, false, false]);
     });
 
     it('should compare array - array', function () {
@@ -160,8 +161,8 @@ describe('smaller', function() {
   describe('DenseMatrix', function () {
 
     it('should compare dense matrix - scalar', function () {
-      assert.deepEqual(smaller('B', matrix(['A', 'B', 'C'])), matrix([false, false, true]));
-      assert.deepEqual(smaller(matrix(['A', 'B', 'C']), 'B'), matrix([true, false, false]));
+      assert.deepEqual(smaller(2, matrix([1, 2, 3])), matrix([false, false, true]));
+      assert.deepEqual(smaller(matrix([1, 2, 3]), 2), matrix([true, false, false]));
     });
 
     it('should compare dense matrix - array', function () {
@@ -180,8 +181,8 @@ describe('smaller', function() {
   describe('SparseMatrix', function () {
 
     it('should compare sparse matrix - scalar', function () {
-      assert.deepEqual(smaller('B', sparse([['A', 'B'], ['C', 'D']])), matrix([[false, false], [true, true]]));
-      assert.deepEqual(smaller(sparse([['A', 'B'], ['C', 'D']]), 'B'), matrix([[true, false], [false, false]]));
+      assert.deepEqual(smaller(2, sparse([[1, 2], [3, 4]])), matrix([[false, false], [true, true]]));
+      assert.deepEqual(smaller(sparse([[1, 2], [3, 4]]), 2), matrix([[true, false], [false, false]]));
     });
 
     it('should compare sparse matrix - array', function () {